backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
yt-dlp==2023.11.16
python-dotenv==1.0.0
pydantic==2.5.2
openai-whisper==20231117
torch==2.1.1
ffmpeg-python==0.2.0 